feat: 방 도메인 DB 구현
- TSID 적용 및 수직 분할을 적용하였다.
- Users 엔티티도 TSID 를 적용하였다.
ISSUE #5

+++ src/main/java/com/tsjeong/brokerage/dto/user/UserReadDto.java
@@ -1,7 +1,7 @@
 package com.tsjeong.brokerage.dto.user;
 
 public record UserReadDto (
-        int id,
+        long id,
         String email,
         String password
-) { }
+) {}
