+++ src/main/java/com/tsjeong/brokerage/repsoitory/UsersRepository.java
@@ -1,12 +1,12 @@
 package com.tsjeong.brokerage.repsoitory;
 
 
-import com.tsjeong.brokerage.entity.Users;
+import com.tsjeong.brokerage.entity.user.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsersRepository extends JpaRepository<Users, Integer> {
+public interface UsersRepository extends JpaRepository<Users, Long> {
 
     Optional<Users> findByEmail(String email);
